:recycle: Refactor a lot of stuff to fix compilation
