+++ cmd/show.go
@@ -44,7 +44,7 @@ This command serves the Qodana report locally and opens a browser to it.`,
 					log.Fatal(err)
 				}
 			} else {
-				core.ShowReport(
+				platform.ShowReport(
 					options.ResultsDir,
 					options.ReportDir,
 					options.Port,
