+++ cmd/view.go
@@ -18,6 +18,7 @@ package cmd
 
 import (
 	"github.com/JetBrains/qodana-cli/v2023/core"
+	"github.com/JetBrains/qodana-cli/v2023/platform"
 	"github.com/spf13/cobra"
 )
 
@@ -38,6 +39,6 @@ func newViewCommand() *cobra.Command {
 		},
 	}
 	flags := cmd.Flags()
-	flags.StringVarP(&options.SarifFile, "sarif-file", "f", core.QodanaSarifName, "Path to the SARIF file")
+	flags.StringVarP(&options.SarifFile, "sarif-file", "f", platform.QodanaSarifName, "Path to the SARIF file")
 	return cmd
 }
