+++ platform/common.go
@@ -2,6 +2,48 @@ package platform
 
 import (
 	"fmt"
+	"github.com/JetBrains/qodana-cli/v2023/cloud"
+	"github.com/pterm/pterm"
+	"github.com/sirupsen/logrus"
+	"net/http"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"runtime"
+	"strings"
+	"time"
+)
+
+//goland:noinspection GoUnnecessarilyExportedIdentifiers
+var (
+	QDJVMC         = "QDJVMC"
+	QDJVM          = "QDJVM"
+	QDAND          = "QDAND"
+	QDPHP          = "QDPHP"
+	QDPY           = "QDPY"
+	QDPYC          = "QDPYC"
+	QDJS           = "QDJS"
+	QDGO           = "QDGO"
+	QDNET          = "QDNET"
+	QDNETC         = "QDNETC"
+	QDANDC         = "QDANDC"
+	QDRST          = "QDRST"
+	QDRUBY         = "QDRUBY"
+	QDCL           = "QDCL"
+	DockerImageMap = map[string]string{
+		QDANDC: "jetbrains/qodana-jvm-android:",
+		QDPHP:  "jetbrains/qodana-php:",
+		QDJS:   "jetbrains/qodana-js:",
+		QDNET:  "jetbrains/qodana-dotnet:",
+		QDNETC: "jetbrains/qodana-cdnet:",
+		QDPY:   "jetbrains/qodana-python:",
+		QDPYC:  "jetbrains/qodana-python-community:",
+		QDGO:   "jetbrains/qodana-go:",
+		QDJVM:  "jetbrains/qodana-jvm:",
+		QDJVMC: "jetbrains/qodana-jvm-community:",
+		QDCL:   "jetbrains/qodana-clang:",
+		//QDRST:  "jetbrains/qodana-rust:",
+	}
 )
 
 // QodanaLogo prepares the info message for the tool
@@ -21,3 +63,228 @@ func QodanaLogo(toolDesc string, version string) string {
 
 `, toolDesc, version)
 }
+
+// GetAnalyzer gets linter for the given path and saves configName
+func GetAnalyzer(path string, yamlName string) string {
+	var analyzers []string
+	PrintProcess(func(_ *pterm.SpinnerPrinter) {
+		languages := readIdeaDir(path)
+		if len(languages) == 0 {
+			languages, _ = recognizeDirLanguages(path)
+		}
+		if len(languages) == 0 {
+			WarningMessage("No technologies detected (no source code files?)\n")
+		} else {
+			WarningMessage("Detected technologies: " + strings.Join(languages, ", ") + "\n")
+			for _, language := range languages {
+				if i, err := langsProductCodes[language]; err {
+					for _, l := range i {
+						analyzers = Append(analyzers, l)
+					}
+				}
+			}
+			if len(analyzers) == 0 {
+				analyzers = AllCodes
+			}
+		}
+	}, "Scanning project", "")
+
+	selector := func(choices []string) string {
+		choice, err := qodanaInteractiveSelect.WithOptions(choices).Show()
+		if err != nil {
+			ErrorMessage("%s", err)
+			return ""
+		}
+		return choice
+	}
+
+	interactive := IsInteractive()
+	analyzer := SelectAnalyzer(path, analyzers, interactive, selector)
+	if analyzer == "" {
+		ErrorMessage("Could not configure project as it is not supported by Qodana")
+		WarningMessage("See https://www.jetbrains.com/help/qodana/supported-technologies.html for more details")
+		os.Exit(1)
+	}
+	SetQodanaLinter(path, analyzer, yamlName)
+	SuccessMessage("Added %s", analyzer)
+	return analyzer
+}
+
+// GetDotNetConfig gets .NET config for the given path and saves configName
+func GetDotNetConfig(projectDir string, yamlName string) bool {
+	possibleOptions := FindFiles(projectDir, []string{".sln", ".csproj", ".vbproj", ".fsproj"})
+	if len(possibleOptions) <= 1 {
+		return false
+	}
+	WarningMessage("Detected multiple .NET solution/project files, select the preferred one \n")
+	choice, err := qodanaInteractiveSelect.WithOptions(possibleOptions).WithDefaultText("Select solution/project").Show()
+	if err != nil {
+		ErrorMessage("%s", err)
+		return false
+	}
+	dotnet := &DotNet{}
+	if strings.HasSuffix(choice, ".sln") {
+		dotnet.Solution = filepath.Base(choice)
+	} else {
+		dotnet.Project = filepath.Base(choice)
+	}
+	return setQodanaDotNet(projectDir, dotnet, yamlName)
+}
+
+// AllNativeCodes is a list of all supported Qodana linters product codes
+var AllNativeCodes = []string{QDNET}
+
+func Image(code string) string {
+	if val, ok := DockerImageMap[code]; ok {
+		return val + releaseVersion
+	} else {
+		logrus.Fatal("Unknown code: " + code)
+		return ""
+	}
+}
+
+func SelectAnalyzer(path string, analyzers []string, interactive bool, selectFunc func([]string) string) string {
+	var analyzer string
+	if len(analyzers) == 0 && !interactive {
+		return ""
+	}
+
+	selection, choices := analyzerToSelect(analyzers, path)
+	logrus.Debugf("Detected products: %s", strings.Join(choices, ", "))
+
+	if len(choices) == 1 || !interactive {
+		analyzer = selection[choices[0]]
+	} else {
+		choice := selectFunc(choices)
+		if choice == "" {
+			return ""
+		}
+		analyzer = selection[choice]
+	}
+	return analyzer
+}
+
+func IsNativeAnalyzer(analyzer string) bool {
+	return Contains(AllNativeCodes, analyzer)
+}
+
+func analyzerToSelect(analyzers []string, path string) (map[string]string, []string) {
+	analyzersMap := make(map[string]string)
+	analyzersList := make([]string, 0, len(analyzers))
+	for _, a := range analyzers {
+		if IsNativeAnalyzer(a) {
+			if IsNativeRequired(path, a) {
+				analyzersMap[a+" (Native)"] = a
+				analyzersList = append(analyzersList, a+" (Native)")
+			}
+		}
+		analyzersMap[Image(a)+" (Docker)"] = Image(a)
+		analyzersList = append(analyzersList, Image(a)+" (Docker)")
+	}
+	return analyzersMap, analyzersList
+}
+
+// ShowReport serves the Qodana report
+func ShowReport(resultsDir string, reportPath string, port int) {
+	cloudUrl := cloud.GetReportUrl(resultsDir)
+	if cloudUrl != "" {
+		openReport(cloudUrl, reportPath, port)
+	} else {
+		WarningMessage("Press Ctrl+C to stop serving the report\n")
+		PrintProcess(
+			func(_ *pterm.SpinnerPrinter) {
+				if _, err := os.Stat(reportPath); os.IsNotExist(err) {
+					logrus.Fatal("Qodana report not found. Get a report by running `qodana scan`")
+				}
+				openReport("", reportPath, port)
+			},
+			fmt.Sprintf("Showing Qodana report from %s", fmt.Sprintf("http://localhost:%d/", port)),
+			"",
+		)
+	}
+}
+
+// openReport serves the report on the given port and opens the browser.
+func openReport(cloudUrl string, path string, port int) {
+	if cloudUrl != "" {
+		resp, err := http.Get(cloudUrl)
+		if err == nil && resp.StatusCode == 200 {
+			err = openBrowser(cloudUrl)
+			if err != nil {
+				return
+			}
+		}
+		return
+	} else {
+		url := fmt.Sprintf("http://localhost:%d", port)
+		go func() {
+			resp, err := http.Get(url)
+			if err == nil && resp.StatusCode == 200 {
+				err := openBrowser(url)
+				if err != nil {
+					return
+				}
+			}
+		}()
+		http.Handle("/", noCache(http.FileServer(http.Dir(path))))
+		err := http.ListenAndServe(fmt.Sprintf(":%d", port), nil)
+		if err != nil {
+			WarningMessage("Problem serving report, %s\n", err.Error())
+			return
+		}
+	}
+	_, _ = fmt.Scan()
+}
+
+// openBrowser opens the default browser to the given url
+func openBrowser(url string) error {
+	var cmd string
+	var args []string
+
+	switch runtime.GOOS {
+	case "windows":
+		cmd = "cmd"
+		args = []string{"/c", "start"}
+	case "darwin":
+		cmd = "open"
+	default: // "linux", "freebsd", "openbsd", "netbsd"
+		cmd = "xdg-open"
+	}
+	args = append(args, url)
+	return exec.Command(cmd, args...).Start()
+}
+
+// noCache handles serving the static files with no cache headers.
+func noCache(h http.Handler) http.Handler {
+	etagHeaders := []string{
+		"ETag",
+		"If-Modified-Since",
+		"If-Match",
+		"If-None-Match",
+		"If-Range",
+		"If-Unmodified-Since",
+	}
+	epoch := time.Unix(0, 0).Format(time.RFC1123)
+	noCacheHeaders := map[string]string{
+		"Expires":         epoch,
+		"Cache-Control":   "no-cache, private, max-age=0",
+		"Pragma":          "no-cache",
+		"X-Accel-Expires": "0",
+	}
+	fn := func(w http.ResponseWriter, r *http.Request) {
+		for _, x := range etagHeaders {
+			if r.Header.Get(x) != "" {
+				r.Header.Del(x)
+			}
+		}
+		for k, v := range noCacheHeaders {
+			w.Header().Set(k, v)
+		}
+		h.ServeHTTP(w, r)
+	}
+	return http.HandlerFunc(fn)
+}
+
+// Version returns the version of the Qodana CLI, set during the GoReleaser build
+var Version = "dev"
+var InterruptChannel chan os.Signal
