+++ platform/native_test.go
@@ -1,4 +1,4 @@
-package core
+package platform
 
 import (
 	"os"
